feat(native-world): complete stores and pools expansion
Atomically extend the FileID layout to 42,341 entries with 8,000 TXDs, 512 COLs and 1,024 IPLs while leaving DAT/path data at its stock 64-entry span. Expand and validate buildings, ColModels and QuadTreeNodes, preserve full-width COL/IPL ownership through MTA consumers, disable stale cache accelerators, and expose pool high-water telemetry.

Add a fail-closed native boundary harness for COL 255/256/511 and IPL 255/256/1023 with exact streaming, side-table, native-pool and MTA dynamic pointer-node rollback. Normalize elevated authorization-file ownership so the ARM64 VM restart gate remains secure and reproducible.

Validated with clang-format, 103 Python tests (2 skipped), the HOODLUM executable validator (1,427 relocation writes and 1,464 total native sites), Release Win32 Game SA/Multiplayer SA/Client Deathmatch/Client Core builds, and user-run Bullworth gameplay. Ticket d831fafb passed all boundary pairs, repeated SA/Bullworth travel, minimize/restore and death/respawn with no overflow, fatal diagnostic or crash.

+++ Client/sdk/game/CColStore.h
@@ -14,7 +14,9 @@
 
 class CVector;
 
-using CollisionSlot = std::uint8_t;
+// GTA's ABI field remains one byte, but the native store runtime preserves the
+// authoritative slot in side storage. Do not truncate it at the MTA boundary.
+using CollisionSlot = std::uint32_t;
 
 class CColStore
 {
